backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose
python-dotenv
email-validator
bcrypt
alembic
